:package: script regarding cookie moved to 'util/cookie' directory
