+++ Frontend/src/pages/bridge/Bridge.tsx
@@ -1,5 +1,6 @@
 import { saveAccessToken, saveLoginInfo, saveRefreshToken } from '@/hooks';
 import { LoginInfo } from '@/types/loginInfo';
+import { deleteCookie, parseCookie } from '@/utils';
 import React from 'react';
 import { useEffect } from 'react';
 import { useNavigate, useLocation } from 'react-router-dom';
@@ -9,25 +10,6 @@ function Bridge() {
   const location = useLocation(); // useLocation 훅을 사용하여 현재 경로 정보를 가져옴
   const queryParams = new URLSearchParams(location.search); // 쿼리 문자열을 파싱하기 위해 사용
 
-  const deleteCookie = (cookieName: string) => {
-    document.cookie = `${cookieName}=; expires=Thu, 01 Jan 1970 00:00:00 UTC; path=/;`;
-  }
-
-  const parseCookie = (cookieString: string) => {
-    //To-Do : cookieString 형식이 맞는지 검증하고 아니면 에러처리?
-    const cookieMap = new Map();
-    cookieString.split(';').map((cookie) => {
-      const [name, value] = cookie.trim().split('=');
-      cookieMap.set(name, value.trim());
-    });
-
-    //To-Do : cookie에서 유효한 값을 찾지 못한 경우 에러처리.
-    const accessToken = cookieMap.get('accessToken');
-    //const refreshToken = cookieMap.get('refreshToken');
-    const memberId = cookieMap.get('memberId');
-    
-    return [accessToken, memberId];
-  }
 
   useEffect(() => {
       // 쿠키를 불러옴
